refactor(routes): consolidate user feed and follow functionality into feed.js
